algo-backend/requirements.txt
fastapi==0.115.6
python-dotenv==1.0.1
pydantic==2.10.3
uvicorn==0.30.6
smartapi-python==1.5.5
pyotp==2.9.0
logzero==1.7.0
pandas==2.2.3
numpy==2.2.4
requests==2.32.5
websocket-client==1.9.0
yfinance==0.2.52
alembic==1.13.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
cryptography==41.0.5
httpx==0.25.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
backtrader==1.9.78.123
